refactor: remove proxy module
