+++ gravitee-apim-console-webui/src/management/api/_api.scss
@@ -4,7 +4,7 @@
 @import 'audit/history/apiHistory';
 @import 'policy-studio-v1/policies/policy';
 @import 'properties-v1/properties';
-@import 'proxy/health-check-dashboard/healthcheck';
+@import 'health-check-dashboard/healthcheck';
 @import 'resources-v1/resources';
 
 .gravitee-api-container {
