+++ gravitee-apim-console-webui/src/management/api/health-check-dashboard/_healthcheck.scss
