+++ gravitee-apim-console-webui/src/management/api/health-check-dashboard/api-proxy-health-check-dashboard.module.ts
