+++ gravitee-apim-console-webui/src/management/api/health-check-dashboard/chart/chart.directive.ts
