+++ gravitee-apim-console-webui/src/management/api/health-check-dashboard/healthcheck-dashboard.component.ajs.ts
