+++ gravitee-apim-console-webui/src/management/api/health-check-dashboard/healthcheck-log.component.ajs.ts
