+++ gravitee-apim-console-webui/src/management/api/health-check-dashboard/healthcheckmetric/healthcheck-metric.component.ts
