+++ gravitee-apim-console-webui/src/management/api/health-check-dashboard/progressbar/progress-bar.html
